Refactor: Move load_image from _stbt.core to _stbt.imgutils
This removes a dependency from _stbt.transition on _stbt.core.
